requirements.txt
playwright==1.42.0
ollama==0.1.7
beautifulsoup4==4.12.3
pydantic==2.6.1
openai
python-dotenv